% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_map.R
\name{export_map}
\alias{export_map}
\title{Export Map}
\usage{
export_map(
  plot,
  filename,
  path = "outputs/maps",
  format = c("pdf"),
  page_size = "A4",
  orientation = "landscape",
  width = NULL,
  height = NULL,
  dpi = 300
)
}
\arguments{
\item{dpi}{}
}
\description{
Export Map
}
