% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosom_leaflet.R
\name{geosom_leaflet}
\alias{geosom_leaflet}
\title{Create an Interactive Leaflet Map of Somalia Administrative Layers}
\usage{
geosom_leaflet(
  layers = c("admin0", "states", "admin1", "admin2"),
  basemap = c("positron", "osm", "esri")
)
}
\arguments{
\item{layers}{Character vector of layers to include. Options are \code{"admin0"},
\code{"states"}, \code{"admin1"}, and \code{"admin2"}.}

\item{basemap}{Character. Basemap to display by default. Options are \code{"osm"},
\code{"positron"}, and \code{"esri"}.}
}
\value{
A \code{leaflet} map object.
}
\description{
Creates an interactive Leaflet map showing Somalia administrative boundary
layers included in the \code{geosom} package. Users can display the national
boundary, state boundaries, administrative level 1 boundaries, and
administrative level 2 boundaries.
}
\examples{
geosom_leaflet()

geosom_leaflet(layers = c("admin1", "admin2"))
}
