% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_geosom_layers.R
\name{list_geosom_layers}
\alias{list_geosom_layers}
\title{List Available geosom Boundary Layers}
\usage{
list_geosom_layers()
}
\value{
A data frame with available layer names and descriptions.
}
\description{
Shows the boundary datasets available in the \code{geosom} package.
}
\examples{
list_geosom_layers()
}
