% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_population.R
\docType{data}
\name{population_admin2_2026}
\alias{population_admin2_2026}
\title{Somalia District-Level Population Estimates for Humanitarian Response Planning, 2026}
\format{
A data frame with 90 rows and 21 variables:
\describe{
\item{admin0_name}{Country name.}
\item{admin1_name}{Administrative level 1 name, usually region.}
\item{admin2_name}{Administrative level 2 name, usually district.}
\item{admin2_pcode}{Administrative level 2 district P-code.}
\item{year}{Population estimate year.}
\item{male_under_5}{Estimated male population under 5 years.}
\item{female_under_5}{Estimated female population under 5 years.}
\item{total_under_5}{Estimated total population under 5 years.}
\item{male_5_17}{Estimated male population aged 5 to 17 years.}
\item{female_5_17}{Estimated female population aged 5 to 17 years.}
\item{total_5_17}{Estimated total population aged 5 to 17 years.}
\item{male_18_59}{Estimated male population aged 18 to 59 years.}
\item{female_18_59}{Estimated female population aged 18 to 59 years.}
\item{total_18_59}{Estimated total population aged 18 to 59 years.}
\item{male_60_plus}{Estimated male population aged 60 years and above.}
\item{female_60_plus}{Estimated female population aged 60 years and above.}
\item{total_60_plus}{Estimated total population aged 60 years and above.}
\item{male_total}{Estimated total male population.}
\item{female_total}{Estimated total female population.}
\item{population_total}{Estimated total district population.}
\item{source}{Source description for the population estimate.}
}
}
\source{
Somalia District-level population estimates for humanitarian response
planning, 2026.
}
\description{
A district-level population dataset for Somalia prepared to support humanitarian
response planning in 2026. The dataset provides population estimates by
administrative level 2 district, disaggregated by sex and broad age groups.
}
\details{
The dataset can be joined with the \code{admin2} boundary layer using the district
P-code field. This supports population mapping, district-level spatial analysis,
humanitarian planning, and reporting workflows.
}
\examples{
data(population_admin2_2026)

head(population_admin2_2026, 10)

admin2 <- get_admin_shape("admin2")

admin2_pop <- dplyr::left_join(
  admin2,
  population_admin2_2026,
  by = c("admin2pcod" = "admin2_pcode")
)

}
\keyword{datasets}
