\name{project3}
\alias{project3}
\docType{data}
\title{Portugal municipal census data 2011}
\description{
  The 2011 Portugal Census data contains the centroid coordinates and 
  11 selected socio-economic variables for the 278 municipalities of Portugal.
  The data is a subset of the complete census data with 50 variables.
}
\usage{data(project3)}
\keyword{datasets}