\name{squareville}
\alias{squareville}
\docType{data}
\title{Small town artificial spatial dataset}
\description{
  Squareville is an artificial dataset purposely created to use in GeoSOM suite.
  Squareville is a small-town with square boundaries and an area of 10000 m2.  
  Squareville has 100 houses evenly spaced with coordinates x in [5, 95] and 
  y in [5, 95]. For each house we know the average salary, which is s in  
  [900 1000] for 35<=x<=65 and s in [0 100]
}
\usage{data(squareville)}
\references{R. Henriques et al., International conference on computational science and its applications (453-466) (2009)}
\keyword{datasets}