% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_shapes.R
\name{get_admin_shape}
\alias{get_admin_shape}
\title{Load Administrative Shapes}
\usage{
get_admin_shape(level = c("admin0", "admin1", "admin2", "states"))
}
\arguments{
\item{level}{specefiy which shapefile to load.
and must be one of `"admin0", "admin1", admin2", "states"`}
}
\value{
An `sf` object containing spatial data
}
\description{
Load Administrative Shapes
}
